Use h2 only on access headings.
When this feature was built, most themes used h2 as primary heading. Now, h1 is much more common, and this is more appropriate. The popup views are mostly in modals, and h2 is appropriate there. In the mini calendar and list calendar popups, the heading structure is different, and this should be h3.

Fixes #515